Refactor MMR calculation

+++ server/evr_match_player.go
@@ -18,7 +18,7 @@ type PlayerInfo struct {
 	RatingMu       float64   `json:"rating_mu,omitempty"`
 	RatingSigma    float64   `json:"rating_sigma,omitempty"`
 	RatingOrdinal  float64   `json:"rating_ordinal,omitempty"`
-	RatingScore    float64   `json:"rating_score,omitempty"`
+	RatingScore    int       `json:"rating_score,omitempty"`
 	Username       string    `json:"username,omitempty"`
 	DiscordID      string    `json:"discord_id,omitempty"`
 	UserID         string    `json:"user_id,omitempty"`
